cambios delete de publicaciones

+++ src/modulos/publicaciones/publicaciones.service.ts
@@ -42,8 +42,16 @@ export class PublicacionesService {
         return proyectoActualizado;
     }
 
-    async removeProyecto(id: number): Promise<void> {
+    async removeProyecto(id: number): Promise<{message: string}> {
+
+        const proyecto = await this.proyectosRepository.findOneBy({ id });
+
+        if (!proyecto) {
+            throw new NotFoundException(`Proyecto con id ${id} no encontrado`);
+        }
         await this.proyectosRepository.delete(id);
+
+        return { message: `Proyecto con id ${id} eliminado correctamente` };
     }
 
     async findAllProyectos(page?: number, limit?: number): Promise<{ data: Proyectos[]; total: number }> {
@@ -96,8 +104,17 @@ export class PublicacionesService {
         return donacionActualizada;
     }
 
-    async removeDonacion(id: number): Promise<void> {
+    async removeDonacion(id: number): Promise<{message: string}> {
+
+        const donacion = await this.donacionesRepository.findOneBy({ id });
+
+        if (!donacion) {
+            throw new NotFoundException(`Donación con id ${id} no encontrada`);
+        }
+
         await this.donacionesRepository.delete(id);
+
+        return { message: `Donación con id ${id} eliminada correctamente` };
     }
 
     async findAllDonacion(page?: number, limit?: number): Promise<{ data: Donacion[]; total: number }> {
@@ -152,8 +169,16 @@ export class PublicacionesService {
         return eventoActualizado;
     }
 
-    async removeEventos(id: number): Promise<void> {
+    async removeEventos(id: number): Promise<{message: string}> {
+
+        const evento = await this.eventosRepository.findOneBy({ id });
+
+        if (!evento) {
+            throw new NotFoundException(`Evento con id ${id} no encontrado`);
+        }
         await this.eventosRepository.delete(id);
+
+        return { message: `Evento con id ${id} eliminado correctamente` };
     }
 
     async findAllEventos(page?: number, limit?: number): Promise<{ data: Eventos[]; total: number }> {
